Randomly subset populations, change mutation technique
Instead of using the entire set of offspring as a population, we take
a subset. Also, the mutation technique has change. Before, genetic
diversity was introduced by breeding the two best solutions with two
random solutions. Now, the two best solutions are bread with one random
solution, and all solutions have a 1 in 10^N chance to mutate N times.
A mutation consists of randomly removing a transition from the solution
and randomly inserting a transition into the solution.